chore: refactor gpt loading logic
