+++ blocks/ad/ad.js
@@ -1,8 +1,36 @@
+import { getJson, loadScript } from '../../scripts/scripts.js';
+
+/**
+ * Retrieves information about the sites ads, which will ultimately be pulled
+ * from the "ads" spreadsheet at the root of the project. There is some caching
+ * in place to ensure fast subsequent retrieval of the ad data in the same
+ * session.
+ * @param {string} adId ID of the ad from the spreadsheet.
+ * @returns Simple object containing all the ad's information, or falsy
+ *  if the ad could not be found.
+ */
+export async function getAd(adId) {
+  const ads = await getJson('/ads.json', 'ads');
+  if (!ads) {
+    return null;
+  }
+  return ads.data.find((c) => c.ID === adId);
+}
+
+function getAdTargets(ad) {
+  if (ad.Targeting) {
+    return String(ad.Targeting).split(',').map((target) => String(target).trim());
+  }
+  return null;
+}
+
 /**
  *
  * @param {HTMLElement} block Ad block to decorate.
  */
-export default function decorate(block) {
+export default async function decorate(block) {
+  window.googletag = window.googletag || { cmd: [] };
+
   if (!block.id) {
     block.id = `ad-${Math.random().toString(32).substring(2)}`;
   }
@@ -11,5 +39,31 @@ export default function decorate(block) {
       block.dataset.adid = String(row.innerText).trim();
     }
     row.remove();
+  });
+
+  const { id } = block;
+  const data = await getAd(block.dataset.adid);
+  const width = parseInt(data.Width, 10);
+  const height = parseInt(data.Height, 10);
+  block.style.width = `${width}px`;
+  block.style.height = `${height}px`;
+  window.googletag.cmd.push(() => {
+    const adSlot = window.googletag
+      .defineSlot(data.Path, [[width, height]], id)
+      .addService(window.googletag.pubads());
+
+    const targets = getAdTargets(data);
+    if (targets) {
+      adSlot.setTargeting(...targets);
+    }
+  });
+  // Enable SRA and services.
+  window.googletag.cmd.push(() => {
+    window.googletag.pubads().enableSingleRequest();
+    window.googletag.enableServices();
+  });
+
+  window.googletag.cmd.push(() => {
+    window.googletag.display(block.id);
   });
 }
